+++ CHANGELOG.md
@@ -12,6 +12,8 @@ FlatLaf Change Log
   - Style classes allow defining style rules at a single place (in UI defaults)
     and use them in any component. (PR #388)\
     E.g.: `mySlider.putClientProperty( "FlatLaf.styleClass", "myclass" );`
+- Native window decorations: Show Windows 11 snap layouts menu when hovering the
+  mouse over the maximize button. (issues #397 and #407)
 - TextField, FormattedTextField and PasswordField: Support leading and trailing
   icons (set client property `JTextField.leadingIcon` or
   `JTextField.trailingIcon` to an `Icon`). (PR #378; issue #368)